pyproject.toml
[project]
name = "ner-accident"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "load-dotenv>=0.1.0",
    "nltk>=3.9.1",
    "pandas>=2.2.3",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
]
